refactor(tools): extract Agentic Memory tools to dedicated package
